ManageAlertViewModel accountId, userId, priceUpdater added

+++ Shared/ViewModels/ManageAlertViewModel.swift
@@ -19,16 +19,22 @@ final class ManageAlertViewModel: ObservableObject {
     
     private var subscriptions: Set<AnyCancellable>
     private let storage: IPriceAlertStorage
+    private let updater: PriceAlertUpdater
+    private let userId: String
+    private let accountId: String
     
-    init(coin: Coin, price: Decimal, alertsStorage: IPriceAlertStorage) {
+    init(accountId: String, userId: String, coin: Coin, price: Decimal, alertsStorage: IPriceAlertStorage) {
+        self.accountId = accountId
+        self.userId = userId
         self.coin = coin
         self.worth = price.formattedString(Currency.fiat(USD))
         self.storage = alertsStorage
-        self.alerts = storage.alerts.reversed()
+        self.alerts = storage.alerts(accountId: accountId, coin: coin.code).reversed()
         self.type = .worthMore(coin)
         self.value = String()
         self.canSet = false
         self.subscriptions = Set<AnyCancellable>()
+        self.updater = PriceAlertUpdater()
         
         subscribe()
     }
@@ -52,27 +58,47 @@ final class ManageAlertViewModel: ObservableObject {
             title = "\(type.description) \(value)%"
         }
         
-        let model = PriceAlertModel(
-            id: UUID().uuidString,
-            coin: coin.code,
-            timestamp: "\(Date().timeIntervalSince1970)",
-            title: title.replacingOccurrences(of: "...", with: ""),
-            value: value
-        )
+        let alertId = UUID().uuidString
+        let coinCode = coin.code
         
-        storage.addAlert(model)
-        
-        reset()
+        updater.createAlert(devId: userId, alertId: alertId, coin: coinCode, price: value) { [unowned self] result in
+            if result.contains("Success") {
+                let model = PriceAlertModel(
+                    accountId: accountId,
+                    id: alertId,
+                    coin: coinCode,
+                    timestamp: "\(Date().timeIntervalSince1970)",
+                    title: title.replacingOccurrences(of: "...", with: ""),
+                    value: value
+                )
+                
+                storage.addAlert(model)
+                
+                reset()
+            } else {
+                print("Cannot crete price alert: \(result)")
+            }
+        }
     }
     
     func deleteAlert(_ alert: PriceAlert) {
-        storage.deleteAlert(alert)
-        alerts = storage.alerts.reversed()
+        guard let alertId = alert.id else { return }
+        
+        updater.deleteAlert(alertId: alertId) { [unowned self] result in
+            if result.contains("Success") {
+                storage.deleteAlert(alert)
+                reset()
+            } else {
+                print("Cannot delete price alert: \(result)")
+            }
+        }
     }
     
     private func reset() {
-        value = String()
-        alerts = storage.alerts.reversed()
+        DispatchQueue.main.async {
+            self.value = String()
+            self.alerts = self.storage.alerts(accountId: self.accountId, coin: self.coin.code).reversed()
+        }
     }
 }
 
@@ -80,7 +106,13 @@ extension ManageAlertViewModel {
     static func config(coin: Coin) -> ManageAlertViewModel {
         let price = Portal.shared.marketDataProvider.ticker(coin: coin)?[.usd].price ?? 1
         let storage = Portal.shared.dbStorage
+        let userId = Portal.shared.userId
+        let accountId = Portal.shared.accountManager.activeAccount?.id
         
-        return ManageAlertViewModel(coin: coin, price: price, alertsStorage: storage)
+        guard let id = accountId else {
+            fatalError("Account id is nil")
+        }
+        
+        return ManageAlertViewModel(accountId: id, userId: userId, coin: coin, price: price, alertsStorage: storage)
     }
 }
